pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "proxy-pool"
version = "2.5.0"
description = "Proxy IP pool system for web scraping"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "gunicorn==22.0.0",
    "lxml>=5.0.0",
    "redis==3.5.3",
    "APScheduler==3.10.4",
    "click==8.0.1",
    "Flask==2.1.1",
    "werkzeug==2.1.0",
    "PyYAML>=6.0",
]

[tool.setuptools]
packages = ["api", "db", "fetcher", "handler", "helper", "log", "util"]
py-modules = ["proxyPool", "setting"]

[project.scripts]
proxy-pool = "proxyPool:cli"